# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Kaspi Pay POS Automation API",
    "version": "1.0.0",
    "description": "REST API for integrating POS systems with Kaspi Pay API (SMS Authorization, Invoicing, QR Payments, Operation History, and Refunds)."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Theater API Key used for authentication and permission gating"
      },
      "AdminBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Master admin secret token check (ADMIN_SECRET_KEY)"
      },
      "TokenSN": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Token-SN",
        "description": "Session token SN obtained after OTP verification"
      },
      "VtokenSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Vtoken-Secret",
        "description": "Encrypted session token secret"
      },
      "ProfileId": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Profile-Id",
        "description": "Organization profile ID (optional, sent after auth)"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Description of the error"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/invoice/client-info": {
      "get": {
        "tags": ["Invoice"],
        "summary": "Get Client Info",
        "description": "Retrieves the name and Kaspi account verification details of a client by their phone number.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "phoneNumber",
            "in": "query",
            "required": true,
            "description": "Customer phone number in 7XXXXXXXXXX format",
            "schema": {
              "type": "string"
            },
            "example": "77001234567"
          }
        ],
        "responses": {
          "200": {
            "description": "Client details loaded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "StatusCode": {
                      "type": "integer",
                      "example": 0
                    },
                    "Data": {
                      "type": "object",
                      "properties": {
                        "FirstName": {
                          "type": "string",
                          "example": "Александр"
                        },
                        "MiddleName": {
                          "type": "string",
                          "example": "Сергеевич"
                        },
                        "LastName": {
                          "type": "string",
                          "example": "П."
                        },
                        "HasKaspiGold": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Phone number required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "phoneNumber required"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API Key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Invalid API Key"
                }
              }
            }
          },
          "412": {
            "description": "Unauthorized / Session error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Invalid or expired vtokenSecret. Re-authenticate."
                }
              }
            }
          },
          "500": {
            "description": "Server or database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Database connection timeout"
                }
              }
            }
          }
        }
      }
    },
    "/api/invoice/create": {
      "post": {
        "tags": ["Invoice"],
        "summary": "Create Remote Invoice",
        "description": "Sends a payment request invoice directly to a customer's phone number.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["phoneNumber", "amount"],
                "properties": {
                  "phoneNumber": {
                    "type": "string",
                    "example": "77001234567"
                  },
                  "amount": {
                    "type": "number",
                    "example": 1000
                  },
                  "comment": {
                    "type": "string",
                    "example": "Оплата заказа #42"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invoice successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "StatusCode": {
                      "type": "integer",
                      "example": 0
                    },
                    "Data": {
                      "type": "object",
                      "properties": {
                        "Id": {
                          "type": "integer",
                          "example": 123456
                        },
                        "Status": {
                          "type": "string",
                          "example": "RemotePaymentCreated"
                        },
                        "Amount": {
                          "type": "number",
                          "example": 1000
                        },
                        "ClientMobile": {
                          "type": "string",
                          "example": "77001234567"
                        },
                        "ReceiptUrl": {
                          "type": "string",
                          "example": "https://..."
                        },
                        "OrderNumber": {
                          "type": "string",
                          "example": "ORDER-123"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid properties",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing API Key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Invalid API Key."
                }
              }
            }
          },
          "403": {
            "description": "API Key suspended, expired, or lacks invoice permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Remote invoicing is not enabled for this API Key."
                }
              }
            }
          },
          "500": {
            "description": "Server or Kaspi API error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Failed to connect to Kaspi API"
                }
              }
            }
          }
        }
      }
    },
    "/api/invoice/details": {
      "get": {
        "tags": ["Invoice"],
        "summary": "Get Invoice Status/Details",
        "description": "Retrieves the status of an existing invoice request using its operation ID.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "operationId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "123456"
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice details returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "StatusCode": {
                      "type": "integer",
                      "example": 0
                    },
                    "Data": {
                      "type": "object",
                      "properties": {
                        "Id": {
                          "type": "integer",
                          "example": 123456
                        },
                        "Amount": {
                          "type": "number",
                          "example": 1000
                        },
                        "Status": {
                          "type": "string",
                          "example": "Processed"
                        },
                        "StatusDesc": {
                          "type": "string",
                          "example": "Оплачено"
                        },
                        "OrderNumber": {
                          "type": "string",
                          "example": "ORDER-12345"
                        },
                        "ClientMobile": {
                          "type": "string",
                          "example": "77071234567"
                        },
                        "ReceiptUrl": {
                          "type": "string",
                          "example": "https://pay.kaspi.kz/receipt/..."
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing operationId parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "operationId query parameter required"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API Key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Invalid API Key"
                }
              }
            }
          },
          "404": {
            "description": "Invoice not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Invoice not found"
                }
              }
            }
          },
          "500": {
            "description": "Server or database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Database connection timeout"
                }
              }
            }
          }
        }
      }
    },
    "/api/invoice/cancel": {
      "post": {
        "tags": ["Invoice"],
        "summary": "Cancel Pending Invoice",
        "description": "Cancels a remote invoice that hasn't been completed yet.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["operationId"],
                "properties": {
                  "operationId": {
                    "type": "string",
                    "example": "123456"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invoice cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "StatusCode": {
                      "type": "integer",
                      "example": 0
                    },
                    "Data": {
                      "type": "object",
                      "properties": {
                        "Status": {
                          "type": "string",
                          "example": "RemotePaymentCanceled"
                        },
                        "StatusDesc": {
                          "type": "string",
                          "example": "Счет отменен"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing operationId parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "operationId required"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API Key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Invalid API Key"
                }
              }
            }
          },
          "404": {
            "description": "Invoice not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Invoice not found"
                }
              }
            }
          },
          "500": {
            "description": "Server or database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Database connection timeout"
                }
              }
            }
          }
        }
      }
    },
    "/api/invoice/history": {
      "post": {
        "tags": ["Invoice"],
        "summary": "Get Recent Invoices History",
        "description": "Loads the history of recent invoices (up to 20).",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invoices history returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "StatusCode": {
                      "type": "integer",
                      "example": 0
                    },
                    "Data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "Id": {
                            "type": "integer",
                            "example": 123456
                          },
                          "Amount": {
                            "type": "number",
                            "example": 1000
                          },
                          "Status": {
                            "type": "string",
                            "example": "Processed"
                          },
                          "StatusDesc": {
                            "type": "string",
                            "example": "Оплачено"
                          },
                          "CreatedDate": {
                            "type": "string",
                            "example": "2026-06-26T16:00:00"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid API Key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Invalid API Key"
                }
              }
            }
          },
          "500": {
            "description": "Server or database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Database connection timeout"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/history/details": {
      "post": {
        "tags": ["History"],
        "summary": "Get Operation Detail",
        "description": "Retrieves precise transaction details (payment types, items, commissions).",
        "security": [
          {
            "TokenSN": [],
            "VtokenSecret": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["id"],
                "properties": {
                  "id": {
                    "type": "integer",
                    "description": "Operation ID",
                    "example": 123456
                  },
                  "operationMethod": {
                    "type": "integer",
                    "example": 0
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Detailed payload returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "StatusCode": {
                      "type": "integer",
                      "example": 0
                    },
                    "Data": {
                      "type": "object",
                      "properties": {
                        "Id": {
                          "type": "integer",
                          "example": 11223344
                        },
                        "Amount": {
                          "type": "number",
                          "example": 5000
                        },
                        "Type": {
                          "type": "string",
                          "example": "Sale"
                        },
                        "CreatedDate": {
                          "type": "string",
                          "example": "2026-06-23T11:45:00.000"
                        },
                        "ConsumerName": {
                          "type": "string",
                          "example": "АЛИХАН Б."
                        },
                        "ConsumerCardMask": {
                          "type": "string",
                          "example": "4405********1122"
                        },
                        "Commission": {
                          "type": "number",
                          "example": 47.5
                        },
                        "NetAmount": {
                          "type": "number",
                          "example": 4952.5
                        },
                        "ReceiptUrl": {
                          "type": "string",
                          "example": "https://pay.kaspi.kz/receipt/..."
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing id parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "id required"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API Key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Invalid API Key"
                }
              }
            }
          },
          "404": {
            "description": "Operation details not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Operation details not found"
                }
              }
            }
          },
          "500": {
            "description": "Server or database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Database connection timeout"
                }
              }
            }
          }
        }
      }
    },
...
  }
}